Fix CI: golangci-lint ineffassign and entry bundle budget
- Remove ineffectual argIdx++ in customfields repo UpdateDefinition
- Extract custom-fields API client and lazy-load user custom fields panel
- Serve customFieldsEnabled via admin-console-capabilities to keep entry chunk lean
- Add 16 B gzip slack to bundle check for Linux CI variance
