+++ clients/web/src/lib/platform-features.ts
@@ -52,7 +52,6 @@ export type PlatformFeaturesSnapshot = {
   impersonationEnabled?: boolean
   bulkCsvImportEnabled?: boolean
   adminSearchEnabled?: boolean
-  customFieldsEnabled?: boolean
   emailTemplateEditorEnabled?: boolean
   maintenanceBannerEnabled?: boolean
   ffZapierConnector?: boolean
@@ -143,7 +142,6 @@ const defaults: PlatformFeaturesSnapshot = {
   impersonationEnabled: false,
   bulkCsvImportEnabled: false,
   adminSearchEnabled: false,
-  customFieldsEnabled: false,
   emailTemplateEditorEnabled: false,
   maintenanceBannerEnabled: true,
   ffZapierConnector: false,
