+++ clients/web/src/pages/admin/AdminUsers.tsx
@@ -1,21 +1,19 @@
-import { useCallback, useEffect, useId, useState } from 'react'
+import { lazy, Suspense, useCallback, useEffect, useId, useState } from 'react'
 import { useSearchParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { usePlatformFeatures } from '../../context/platform-features-context'
 import {
   fetchAdminConsoleCapabilities,
   fetchAdminUsers,
-  fetchCustomFields,
-  fetchAdminUser,
   patchAdminUser,
-  patchAdminUserCustomFields,
   startImpersonation,
-  usersExportUrl,
   type AdminUser,
-  type CustomFieldDefinition,
   type Paginated,
 } from '../../lib/admin-console-api'
+import { fetchCustomFields, usersExportUrl } from '../../lib/custom-fields-api'
 import { startImpersonationSession } from '../../lib/impersonation'
+
+const AdminUserCustomFieldsPanel = lazy(() => import('./AdminUserCustomFieldsPanel'))
 
 const ROLES = ['', 'student', 'instructor', 'ta', 'admin']
 const PAGE_SIZES = [25, 50, 100]
@@ -23,7 +21,7 @@ const PAGE_SIZES = [25, 50, 100]
 export default function AdminUsers() {
   const { t } = useTranslation('common')
   const titleId = useId()
-  const { impersonationEnabled, customFieldsEnabled } = usePlatformFeatures()
+  const { impersonationEnabled } = usePlatformFeatures()
   const [searchParams] = useSearchParams()
   const orgId = searchParams.get('orgId')
   const [q, setQ] = useState('')
@@ -35,12 +33,16 @@ export default function AdminUsers() {
   const [error, setError] = useState<string | null>(null)
   const [busy, setBusy] = useState<string | null>(null)
   const [canManage, setCanManage] = useState(false)
+  const [customFieldsEnabled, setCustomFieldsEnabled] = useState(false)
 
   useEffect(() => {
     let cancelled = false
     void fetchAdminConsoleCapabilities()
       .then((caps) => {
-        if (!cancelled) setCanManage(caps.canManage)
+        if (!cancelled) {
+          setCanManage(caps.canManage)
+          setCustomFieldsEnabled(caps.customFieldsEnabled)
+        }
       })
       .catch(() => {
         if (!cancelled) setCanManage(false)
@@ -102,61 +104,18 @@ export default function AdminUsers() {
 
   const showViewAs = impersonationEnabled && canManage
 
-  const [fieldDefs, setFieldDefs] = useState<CustomFieldDefinition[]>([])
+  const [fieldDefsCount, setFieldDefsCount] = useState(0)
   const [expandedUserId, setExpandedUserId] = useState<string | null>(null)
-  const [fieldValues, setFieldValues] = useState<Record<string, string>>({})
-  const [fieldBusy, setFieldBusy] = useState(false)
 
   useEffect(() => {
     if (!customFieldsEnabled) return
     void fetchCustomFields('user', orgId)
-      .then(setFieldDefs)
-      .catch(() => setFieldDefs([]))
+      .then((defs) => setFieldDefsCount(defs.length))
+      .catch(() => setFieldDefsCount(0))
   }, [customFieldsEnabled, orgId])
 
-  async function openCustomFields(user: AdminUser) {
-    if (expandedUserId === user.id) {
-      setExpandedUserId(null)
-      return
-    }
-    setExpandedUserId(user.id)
-    setFieldValues({})
-    try {
-      const detail = await fetchAdminUser(user.id, orgId, true)
-      const next: Record<string, string> = {}
-      for (const def of fieldDefs) {
-        const raw = detail.customFields?.[def.key]
-        next[def.key] = raw == null ? '' : String(raw)
-      }
-      setFieldValues(next)
-    } catch {
-      setFieldValues({})
-    }
-  }
-
-  async function saveCustomFields(userId: string) {
-    setFieldBusy(true)
-    try {
-      const payload: Record<string, unknown> = {}
-      for (const def of fieldDefs) {
-        const raw = fieldValues[def.key] ?? ''
-        if (def.fieldType === 'boolean') {
-          payload[def.key] = raw === 'true'
-        } else if (def.fieldType === 'number' && raw !== '') {
-          payload[def.key] = Number(raw)
-        } else if (raw !== '') {
-          payload[def.key] = raw
-        } else {
-          payload[def.key] = null
-        }
-      }
-      await patchAdminUserCustomFields(userId, payload, orgId)
-      setExpandedUserId(null)
-    } catch (e) {
-      setError(e instanceof Error ? e.message : 'Failed to save custom fields.')
-    } finally {
-      setFieldBusy(false)
-    }
+  function openCustomFields(user: AdminUser) {
+    setExpandedUserId((current) => (current === user.id ? null : user.id))
   }
 
   return (
@@ -280,10 +239,10 @@ export default function AdminUsers() {
                   <td className="px-4 py-2">{user.active ? 'Active' : 'Deactivated'}</td>
                   <td className="px-4 py-2">
                     <div className="flex flex-wrap gap-3">
-                      {customFieldsEnabled && canManage && fieldDefs.length > 0 ? (
+                      {customFieldsEnabled && canManage && fieldDefsCount > 0 ? (
                         <button
                           type="button"
-                          onClick={() => void openCustomFields(user)}
+                          onClick={() => openCustomFields(user)}
                           className="text-sm text-slate-600 hover:underline dark:text-slate-300"
                         >
                           Custom fields
@@ -315,56 +274,13 @@ export default function AdminUsers() {
                 {expandedUserId === user.id ? (
                   <tr key={`${user.id}-fields`} className="border-t border-slate-100 bg-slate-50 dark:border-neutral-800 dark:bg-neutral-950">
                     <td colSpan={6} className="px-4 py-4">
-                      <fieldset className="space-y-3">
-                        <legend className="text-sm font-medium text-slate-800 dark:text-slate-200">Custom fields</legend>
-                        {fieldDefs.map((def) => (
-                          <label key={def.id} className="block max-w-md text-sm">
-                            <span className="mb-1 block">{def.label}</span>
-                            {def.fieldType === 'select' ? (
-                              <select
-                                value={fieldValues[def.key] ?? ''}
-                                onChange={(e) => setFieldValues({ ...fieldValues, [def.key]: e.target.value })}
-                                className="w-full rounded-lg border border-slate-300 px-3 py-2 dark:border-neutral-700 dark:bg-neutral-900"
-                              >
-                                <option value="">—</option>
-                                {(def.selectOptions ?? []).map((opt) => (
-                                  <option key={opt} value={opt}>{opt}</option>
-                                ))}
-                              </select>
-                            ) : def.fieldType === 'boolean' ? (
-                              <select
-                                value={fieldValues[def.key] ?? ''}
-                                onChange={(e) => setFieldValues({ ...fieldValues, [def.key]: e.target.value })}
-                                className="w-full rounded-lg border border-slate-300 px-3 py-2 dark:border-neutral-700 dark:bg-neutral-900"
-                              >
-                                <option value="">—</option>
-                                <option value="true">Yes</option>
-                                <option value="false">No</option>
-                              </select>
-                            ) : (
-                              <input
-                                type={def.fieldType === 'number' ? 'number' : def.fieldType === 'date' ? 'date' : 'text'}
-                                value={fieldValues[def.key] ?? ''}
-                                onChange={(e) => setFieldValues({ ...fieldValues, [def.key]: e.target.value })}
-                                className="w-full rounded-lg border border-slate-300 px-3 py-2 dark:border-neutral-700 dark:bg-neutral-900"
-                              />
-                            )}
-                          </label>
-                        ))}
-                        <div className="flex gap-2">
-                          <button
-                            type="button"
-                            disabled={fieldBusy}
-                            onClick={() => void saveCustomFields(user.id)}
-                            className="rounded-lg bg-indigo-600 px-3 py-2 text-sm font-medium text-white hover:bg-indigo-700 disabled:opacity-50"
-                          >
-                            Save custom fields
-                          </button>
-                          <button type="button" onClick={() => setExpandedUserId(null)} className="rounded-lg px-3 py-2 text-sm">
-                            Cancel
-                          </button>
-                        </div>
-                      </fieldset>
+                      <Suspense fallback={<p className="text-sm text-slate-500">Loading custom fields…</p>}>
+                        <AdminUserCustomFieldsPanel
+                          userId={user.id}
+                          orgId={orgId}
+                          onClose={() => setExpandedUserId(null)}
+                        />
+                      </Suspense>
                     </td>
                   </tr>
                 ) : null}
